+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Added
 
+- Add --skip-changelog to release without skipping pre-release gates or the default task.
+
 ### Changed
 
 ### Deprecated